用户储存修改为二进制

+++ app/src/main/java/com/xyf/lockers/listener/OnAllLockersStatusListener.java
@@ -10,6 +10,6 @@ package com.xyf.lockers.listener;
  */
 public interface OnAllLockersStatusListener extends BaseSerialportInterface {
 
-    void onAllLockersStatusResponse(int allLockers);
+    void onAllLockersStatusResponse(long allLockers);
 
 }
